React 3 fix
