+++ RomanKaluger/Messenger/src/components/messenger/ChatsOnDialogPanel/index.js
@@ -1,0 +1,1 @@
+export ChatsOnDialogPanel from './ChatsOnDialogPanel';
