+++ RomanKaluger/Messenger/src/components/messenger/ContactsDialogPanel/UserSelector.jsx
@@ -1,18 +1,12 @@
-import React, { useCallback, useRef} from "react";
-import { useSelector } from "react-redux";
-import {activateBtn, disableBtn} from "../../utils/helpers";
-import {DbWorker} from "../../utils/DbWorker";
+import React, {useCallback, useRef} from "react";
+import {activateBtn, disableBtn} from "../../../utils/helpers";
+import {DbWorker} from "../../../utils/DbWorker";
 import ListItem from "@material-ui/core/ListItem";
-import {setSelectedChat} from "../../redux/actions";
 import Grid from "@material-ui/core/Grid";
 import Divider from "@material-ui/core/Divider";
-import List from "@material-ui/core/List";
+import {userTypes} from "./userTypes";
 
-const userTypes = {
-    FRIEND: 'FRIEND',
-    OTHER: 'OTHER'
-};
-function UserSelector({user, type, clearInput}) {
+export default function ({user, type, clearInput}) {
     const startChatRef = useRef();
     const startChatHandler = useCallback(async () => {
         clearInput();
@@ -48,40 +42,3 @@ function UserSelector({user, type, clearInput}) {
         </>
     );
 }
-export default ({users, clearInput}) => {
-    const user = useSelector(s => s.app.curUser);
-    const friends = [];
-    const others = [];
-    users.forEach(u => {
-        if (user.friends.includes(u._id)){
-            friends.push(u);
-        } else {
-            others.push(u);
-        }
-    });
-    return(
-        <>
-            <div className={'ChatsSectionDlgPn'}>
-                <div className={'ChatsSectionHeaderDlgPn'}>
-                    Друзья
-                </div>
-                {
-                    friends.length ? <List>{users.map(u => <UserSelector key={u._id} type={userTypes.FRIEND} user={u}/>)}</List> : 'Друзей нет'
-                }
-
-            </div>
-            {
-                others.length ?
-                    <div className={'ChatsSectionDlgPn'}>
-                        <div className={'ChatsSectionHeaderDlgPn'}>
-                            Поиск
-                        </div>
-                        <List>
-                        {others.map(u => <UserSelector clearInput={clearInput} key={u._id}  type={userTypes.OTHER} user={u}/>)}
-                        </List>
-                    </div>
-                    : null
-            }
-        </>
-    );
-};
