+++ RomanKaluger/Messenger/src/components/messenger/ContactsDialogPanel/userTypes.js
@@ -1,0 +1,4 @@
+export const userTypes = {
+    FRIEND: 'FRIEND',
+    OTHER: 'OTHER'
+};
