+++ RomanKaluger/Messenger/src/components/messenger/MessagesPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import MessengerArea from "./MessengerArea";
-import SelectDialogPanel from "./SelectDialogPanel";
+import SelectDialogPanel from "./SelectDialogPanel/SelectDialogPanel";
 
 export default () => {
     return (
