+++ RomanKaluger/Messenger/src/components/messenger/SelectDialogPanel/SelectDialogPanel.jsx
@@ -1,54 +1,19 @@
 import React, { useRef, useCallback, useState, useEffect} from "react";
 import SearchMessPanel from "./SearchMessPanel";
 import {useDispatch, useSelector} from "react-redux";
-import connectionConfig from "../../configs/connectionConfig";
-import ChatsOnDialogPanel from "./ChatsOnDialogPanel";
-import ContactsOnDialogPanel from "./ContactsOnDialogPanel";
-import {setSelectedChat} from "../../redux/actions";
+import connectionConfig from "../../../configs/connectionConfig";
+import ChatsOnDialogPanel from "../ChatsOnDialogPanel/ChatsOnDialogPanel";
+import ContactsOnDialogPanel from "../ContactsDialogPanel/ContactsOnDialogPanel";
 import ShortMenu from "./ShortMenu";
 import classNames from 'classnames';
-import {DbWorker} from "../../utils/DbWorker";
-const categories = {
-    CHATS: 'CHATS',
-    PEOPLE: 'PEOPLE'
-};
+import find from "./findChats";
+import {categories} from "./categories";
+
 const initialState = {
     contacts: [],
     chats: [],
     category: categories.CHATS
 };
-
-async function find(input, curUser, setDialogs, selectedChat, dispatch) {
-    try {
-        if (curUser) {
-            let defaultChatsRes, defaultContactsRes;
-            if (!input) {
-                defaultChatsRes = await DbWorker.authGet(`${connectionConfig.hostHttp}/chats/owner/${curUser._id}`, curUser);
-                defaultContactsRes = await DbWorker.authGet(`${connectionConfig.hostHttp}/users/user/friends/ownerid/${curUser._id}`, curUser);
-            } else {
-                defaultChatsRes = await DbWorker.authGet(`${connectionConfig.hostHttp}/chats/title/${curUser._id}/${input}`, curUser);
-                defaultContactsRes = await DbWorker.authGet(`${connectionConfig.hostHttp}/users/name/${input}`, curUser);
-            }
-            const defaultContacts = await defaultContactsRes.json();
-            const defaultChats = await defaultChatsRes.json();
-            setDialogs(prev => ({
-                contacts: defaultContacts,
-                chats: defaultChats,
-                category: prev.category
-            }));
-            if (selectedChat) {
-                const newSelChatVersion = defaultChats.find(ch => ch.sharedId === selectedChat.sharedId) || null;
-                dispatch(setSelectedChat(newSelChatVersion));
-            }
-        }
-    } catch (e) {
-        setDialogs({
-            contacts: [],
-            chats: [],
-            category: categories.CHATS
-        });
-    }
-}
 
 export default () => {
     const [dialogs, setDialogs] = useState(initialState);
